refactor(core): convert Cardano/util and Cardano/StakePool into dirs
