+++ packages/core/src/Cardano/types/StakePool/index.ts
@@ -1,0 +1,3 @@
+export * from './StakePool';
+export * from './ExtendedStakePoolMetadata';
+export * from './primitives';
