+++ packages/core/src/Cardano/types/StakePool/primitives.ts
@@ -1,5 +1,5 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import * as util from '../util';
+import * as util from '../../util';
 
 /**
  * bech32-encoded pool operator verification key hash
@@ -8,7 +8,7 @@ export type PoolId = util.OpaqueString<'PoolId'>;
 
 /**
  * @param {string} value blake2b_224 digest of an operator verification key hash
- * @throws {util.InvalidStringError}
+ * @throws InvalidStringError
  */
 export const PoolId = (value: string): PoolId => {
   util.assertIsBech32WithPrefix(value, 'pool', 45);
@@ -19,7 +19,7 @@ export type PoolIdHex = util.OpaqueString<'PoolIdHex'>;
 
 /**
  * @param {string} value operator verification key hash as hex string
- * @throws {util.InvalidStringError}
+ * @throws InvalidStringError
  */
 export const PoolIdHex = (value: string): PoolIdHex => {
   util.assertIsHexString(value, 56);
