--- packages/core/src/Cardano/types/Strings/Hashes/index.ts
@@ -1,1 +1,0 @@
-export * from './StakePool';
