+++ packages/core/src/Cardano/util/computeMinUtxoValue.ts
@@ -1,0 +1,4 @@
+/**
+ * Blockchain restriction for minimum coin quantity in a UTxO
+ */
+export const computeMinUtxoValue = (coinsPerUtxoWord: bigint): bigint => coinsPerUtxoWord * 29n;
