+++ packages/core/test/Cardano/util/computeMinUtxoValue.test.ts
@@ -1,0 +1,5 @@
+import { Cardano } from '../../../src';
+
+describe('Cardano.util.computeMinUtxoValue', () => {
+  it('returns bigint', () => expect(typeof Cardano.util.computeMinUtxoValue(100n)).toBe('bigint'));
+});
